feat: move balances to check
